+++ example/android/app/src/main/kotlin/com/bridges/star_micronics_print_example/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.bridges.star_micronics_print_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}